+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Snippets: Updated all snippets to use `v0.28.0` schema
 - Commands: Removed configuration commands from editor
+- Detection: Improved detection of Dev Proxy processes
 
 ### Fixed: